+++ nvim/colors/dracula.lua
@@ -112,24 +112,12 @@ hl('Ignore', { fg = colors.comment })
 hl('Error', { fg = colors.red, bold = true })
 hl('Todo', { fg = colors.pink, bold = true })
 
--- Diff
-hl('DiffAdd', { fg = colors.green, bg = colors.line_highlight })
-hl('DiffChange', { fg = colors.yellow, bg = colors.line_highlight })
-hl('DiffDelete', { fg = colors.red, bg = colors.line_highlight })
-hl('DiffText', { fg = colors.cyan, bg = colors.line_highlight, bold = true })
-
 -- Git signs
 hl('GitSignsAdd', { fg = colors.green })
 hl('GitSignsChange', { fg = colors.yellow })
 hl('GitSignsDelete', { fg = colors.red })
 hl('GitSignsTopdelete', { fg = colors.red })
 hl('GitSignsChangedelete', { fg = colors.orange or colors.yellow })
-hl('GitSignsAddLn', { bg = '#324847' })
-hl('GitSignsChangeLn', { bg = '#454849' })
-hl('GitSignsDeleteLn', { bg = '#473442' })
-hl('GitSignsAddNr', { fg = colors.green, bg = '#324847' })
-hl('GitSignsChangeNr', { fg = colors.yellow, bg = '#454849' })
-hl('GitSignsDeleteNr', { fg = colors.red, bg = '#473442' })
 
 -- Diagnostics
 hl('DiagnosticError', { fg = colors.red })